deleted testing prints;moved EarlyStop to utils;simplier the func __str__;deleted the space
