+++ QCircLearning/utils.py
@@ -5,3 +5,25 @@ def data_augment(data, n_points, shift):
     for _ in range(n_points):
         new_data.append(data + np.random.choice([-1, 0, 1], len(data)) * shift)
     return new_data
+
+class EarlyStopping:
+    def __init__(self, patience=5, min_delta=0.0, verbose=False):
+        self.patience = patience
+        self.min_delta = min_delta
+        self.verbose = verbose
+        self.best_loss = None
+        self.counter = 0
+        self.early_stop = False
+
+    def __call__(self, current_loss):
+        if self.best_loss is None:
+            self.best_loss = current_loss
+        elif current_loss < self.best_loss - self.min_delta:
+            self.best_loss = current_loss
+            self.counter = 0
+        else:
+            self.counter += 1
+            if self.verbose:
+                print(f"EarlyStopping counter: {self.counter}/{self.patience}")
+            if self.counter >= self.patience:
+                self.early_stop = True
